module3-task1: + offer-type enum
