+++ src/common/enums/offer/offer-city.enum.js
@@ -1,0 +1,10 @@
+const OfferCity = {
+  PARIS: `Paris`,
+  COLOGNE: `Cologne`,
+  BRUSSELS: `Brussels`,
+  AMSTERDAM: `Amsterdam`,
+  HAMBURG: `Hamburg`,
+  DUSSELDORF: `Dusseldorf`,
+};
+
+export {OfferCity};
