+++ src/common/enums/offer/offer-type.enum.js
@@ -1,0 +1,8 @@
+const OfferType = {
+  APARTMENT: `apartment`,
+  ROOM: `room`,
+  HOUSE: `house`,
+  HOTEL: `hotel`,
+};
+
+export {OfferType};
