+++ src/common/enums/offer/offers.js
@@ -1,10 +1,2 @@
-const OfferCity = {
-  PARIS: `Paris`,
-  COLOGNE: `Cologne`,
-  BRUSSELS: `Brussels`,
-  AMSTERDAM: `Amsterdam`,
-  HAMBURG: `Hamburg`,
-  DUSSELDORF: `Dusseldorf`,
-};
-
-export {OfferCity};
+export * from './offer-city.enum';
+export * from './offer-type.enum';
